% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{af}
\alias{af}
\title{Acquisition function, selects $n_2$ units from the test set}
\usage{
af(
  data,
  tau_train,
  tau_test,
  theta,
  n2,
  type = "emcite",
  weight_types = 5,
  B = 2
)
}
\arguments{
\item{data}{list containing train and test covariates}

\item{tau_train}{estimated ITE for train set}

\item{tau_test}{predicted ITE for test set}

\item{theta}{initialization of theta}

\item{n2}{number of samples to select}

\item{type}{which method to run: random/variance/type-s/emcite}

\item{weight_types}{weight of type-s error in EMCITE}

\item{B}{number of bootstrap predictions in EMCITE}
}
\value{
vector of selected indexes
}
\description{
Acquisition function, selects $n_2$ units from the test set
}
