% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{dgp}
\alias{dgp}
\title{Create data according to simulation setting}
\usage{
dgp(lista)
}
\arguments{
\item{lista}{A list with the setup functions of the DGP}
}
\value{
A list of the data, for X, y0, y1, yobs, z, tau
}
\description{
Create data according to simulation setting
}
\examples{
dgp(list("N"=1231,"p"=4, "covariate"="linear", "y_mean"="linear", "ite"="linear")
}
