% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_gradient_descent}
\alias{fit_gradient_descent}
\title{Fit stochastic gradient descent}
\usage{
fit_gradient_descent(X, tau_predictions, weight_types)
}
\arguments{
\item{X}{the feature matrix}

\item{weight_types}{value for $\alpha$}

\item{tau_prediction}{ITE predictions}
}
\value{
vector of weights
}
\description{
Fit stochastic gradient descent
}
