% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{predict_ite}
\alias{predict_ite}
\title{Predict ITE from trained BART model on given data}
\usage{
predict_ite(data, models)
}
\arguments{
\item{data}{data to predict on}

\item{models}{models that were trained previously}
}
\value{
predictions of y1, y0, tau and propensity score
}
\description{
Predict ITE from trained BART model on given data
}
