% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{retrain_and_metrics}
\alias{retrain_and_metrics}
\title{Retrain and evaluate the model based on PEHE}
\usage{
retrain_and_metrics(data)
}
\arguments{
\item{data}{data that contains the sampling phase}
}
\description{
Retrain and evaluate the model based on PEHE
}
