% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{sampling_data}
\alias{sampling_data}
\title{Create n_1+n_2 dataset}
\usage{
sampling_data(data, indexes)
}
\arguments{
\item{data}{data to add together}

\item{indexes}{indexes to move from rollout to sampling}
}
\description{
Create n_1+n_2 dataset
}
