% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{sgd_py}
\alias{sgd_py}
\title{Run model in python}
\usage{
sgd_py(X, y, w, theta = NULL)
}
\arguments{
\item{X}{feature matrix}

\item{y}{outcomes}

\item{w}{weight vector}

\item{theta}{initial parameter vector}
}
\description{
Run model in python
}
