% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgp.R
\name{split_data}
\alias{split_data}
\title{Split the data into train and test set}
\usage{
split_data(data, n1)
}
\arguments{
\item{data}{data to split to experiment-roll-out}

\item{n1}{number of units in experimentation phase}
}
\description{
Split the data into train and test set
}
