% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{train_model}
\alias{train_model}
\title{Train BART model for ITE estimation}
\usage{
train_model(data)
}
\arguments{
\item{data}{Data to train on, previously splitted into train/test set}
}
\value{
outcome Outcome psBART

ps propensity score BART model
}
\description{
Train BART model for ITE estimation
}
